requirements.txt
asn1crypto==0.24.0
astroid==1.6.4
autopep8==1.3.5
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.0.3
coveralls==1.3.0
cryptography==1.9
docopt==0.6.2
Flask==1.0.2
Flask-Script==2.0.6
gunicorn==19.8.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
psycopg2==2.7.4
pycodestyle==2.4.0
pycparser==2.18
PyJWT==1.6.4
pylint==1.9.1
python-coveralls==2.9.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
